// src/types/api/clientEvaluation.ts
/**
 * LifeMapWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import {Household} from "./household";


export interface ClientEvaluation { 
    ClientEvaluationID?: number;
    Description?: string;
    ClientID?: number;
    ComplexityOfNeedsID?: number | string;
    NeedsDetail?: string;
    GoalsDetail?: string;
    ConcernsDetail?: string;
    LegacyInterestID?: number | string;
    LegacyDetail?: string;
    CreationDate?: string;
    CreatedBy?: number | string;
    LastModifiedDate?: string;
    LastModifiedBy?: number;
    ComplexityOfNeedsScore: number;
    LegacyInterestScore: number;

    HouseholdID?: number;
    Household: Household;
}